requirements.txt
aiohttp==3.8.3
guilded.py==1.8.0
psycopg2==2.9.3
asyncio==3.4.3
quart==0.18.3
quart-cors==0.6.0
psycopg==3.1.4
psycopg-pool==3.1.4
psycopg-binary==3.1.4
requests==2.27.1
regex==2022.10.31
hypercorn==0.14.3
simplejson==3.18.3